+++ source/parlex/include/parlex/detail/grammar.hpp
@@ -44,25 +44,25 @@ public:
 	state_machine_base const& get_root_state_machine() const override;
 	std::vector<state_machine_base const *> get_state_machines() const override;
 	std::vector<recognizer const *> const & get_recognizers() const;
-	size_t get_recognizer_count() const override;
-	recognizer const& get_recognizer(size_t const index) const override;
-	bool does_precede(size_t const lhs, size_t const rhs) const override;
+	uint16_t get_recognizer_count() const override;
+	recognizer const& get_recognizer(uint16_t const index) const override;
+	bool does_precede(uint16_t const lhs, uint16_t const rhs) const override;
 	precedence_collection get_precedences() const override;
 
-	size_t lookup_production_local_index(std::string const & name) const;
-	size_t lookup_literal_recognizer_index(std::u32string const & content) const;
-	size_t lookup_recognizer_index(std::string const & name) const;
-	size_t lookup_recognizer_index(recognizer const & recognizer) const override;
+	uint16_t lookup_production_local_index(std::string const & name) const;
+	uint16_t lookup_literal_recognizer_index(std::u32string const & content) const;
+	uint16_t lookup_recognizer_index(std::string const & name) const;
+	uint16_t lookup_recognizer_index(recognizer const & recognizer) const override;
 private:
 	std::vector<production> local_productions;
 	std::vector<string_terminal> local_literals;
 	std::vector<recognizer const *> recognizers; // pointers into builtin_terminals, local production, and local literals. The index is used to uniquely identify the corresponding recognizer throughtout the parsing algorithms
-	std::map<recognizer const *, size_t> recognizer_ptr_to_recognizer_index;
-	std::map<std::u32string, size_t> content_to_recognizer_index;
-	std::map<std::u32string, size_t> content_to_local_literal_index;
-	std::map<std::string, size_t> name_to_recognizer_index;
-	std::map<std::string, size_t> name_to_local_production_index;
-	size_t root_production_index;
+	std::map<recognizer const *, uint16_t> recognizer_ptr_to_recognizer_index;
+	std::map<std::u32string, uint16_t> content_to_recognizer_index;
+	std::map<std::u32string, uint16_t> content_to_local_literal_index;
+	std::map<std::string, uint16_t> name_to_recognizer_index;
+	std::map<std::string, uint16_t> name_to_local_production_index;
+	uint16_t root_production_index;
 	precedence_collection precedences;
 
 	size_t add_table_data(std::map<std::string, recognizer const *> & nameToRecognizerPtr, recognizer const * recognizerPtr);
