+++ source/parlex/src/match.cpp
@@ -5,7 +5,7 @@
 namespace parlex {
 namespace detail {
 
-match::match(size_t documentPosition, size_t recognizerIndex, size_t const consumedCharacterCount) : match_class(documentPosition, recognizerIndex), consumed_character_count(consumedCharacterCount) {}
+match::match(uint32_t documentPosition, uint16_t recognizerIndex, uint32_t const consumedCharacterCount) : match_class(documentPosition, recognizerIndex), consumed_character_count(consumedCharacterCount) {}
 match::match(struct match_class const & matchClass, size_t const consumedCharacterCount) : match_class(matchClass), consumed_character_count(consumedCharacterCount) { }
 
 match & match::operator=(match && move) noexcept {
