+++ source/parlex/src/match_class.cpp
@@ -3,7 +3,7 @@
 namespace parlex {
 namespace detail {
 
-match_class::match_class(size_t const documentPosition, size_t const recognizerIndex) : document_position(documentPosition), recognizer_index(recognizerIndex) {
+match_class::match_class(uint32_t const documentPosition, uint16_t const recognizerIndex) : document_position(documentPosition), recognizer_index(recognizerIndex) {
 }
 
 bool match_class::operator <(match_class const & rhs) const {
