+++ source/parlex/src/unicode.cpp
@@ -1,7 +1,7 @@
 #include "parlex/detail/unicode.hpp"
 
 #define U32(name, condition) \
-bool parlex::detail::name##_t::test(std::u32string const& document, size_t documentPosition) const { \
+bool parlex::detail::name##_t::test(std::u32string const& document, uint32_t documentPosition) const { \
 	auto const & i = document[documentPosition]; \
 	return condition; \
 }
